pyproject.toml
[tool.poetry]
name = "themis-users"
version = "0.1.0"
description = ""
authors = ["Ramos <andfera201@gmail.com>"]
readme = "README.md"
packages = [{ include = "*", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.1"
tortoise-orm = "^0.25.1"
asyncpg = "^0.30.0"
pydantic = "^2.11.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pre-commit = "^4.2.0"
radon = "^6.0.1"
xenon = "^0.9.3"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
